+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Removed:
 
 Improved/Fixed:
 
+- fixed currentContext marker in vscode mode
 ## v5.9.1 2026-03-13
 
 Added: